thêm chức năng tạo bài tập cho người hoc "song song tạo chức năng tạo bài tập "admin" và "mentor"

+++ src/main/java/com/aesp/repository/IUserRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface IUserRepository {
     
-    Optional<User> findById(int id);
+    Optional<User> findById(Long id);
     
     List<User> findAll();
 
@@ -26,8 +26,9 @@ public interface IUserRepository {
 
     User updateUser(User user);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
 	Optional<User> findByEmail(String email);
+
 
 }
